WIP new cli
