+++ setup.py
@@ -25,8 +25,9 @@ setup(
     ],
     packages=find_packages(exclude=("tests",)),
     include_package_data=True,
-    install_requires=["pynecone==0.0.61",
-                      "requests",
+    install_requires=["pynecone==0.0.69",
+                      "requests==2.23.0",
+                      "paho-mqtt",
                       "python-dotenv",
                       "tabulate"],
     entry_points={
